// gestao-integrada-rh-source/src/types.ts
export type PerfilUsuario = 'CANDIDATO' | 'FUNCIONARIO' | 'RH_ADMIN';
// ...
export interface Usuario {
  id: string;
  empresaId: string;
  nome: string;
  email: string;
  senhaHash?: string;
  perfil: PerfilUsuario;
  ativo: boolean;
  cargo?: string;
  departamento?: string;
  salarioBase?: number;
  dataAdmissao?: string;
  fotoUrl?: string;
  criadoEm: string;
  atualizadoEm: string;
}